+++ client/src/mock-data/mockUser.jsx
@@ -2,6 +2,7 @@ const mockUser = {
   userID: 1,
   name: "John Doe",
   email: "johndoe@gmail.com",
+  isAdmin: true,
 };
 
 export default mockUser;
