+++ client/src/prop-types/UserType.jsx
@@ -4,6 +4,7 @@ const UserType = PropTypes.shape({
   userID: PropTypes.number.isRequired,
   name: PropTypes.string.isRequired,
   email: PropTypes.string.isRequired,
+  isAdmin: PropTypes.boolean.isRequired,
   // eslint-disable-next-line
 });
